Made a recipe object
The way I was going about things was totally
different from the solution. I was attempting
to nest objects inside of other objects in the
hopes of having an array filled with the
ingrediet quantity entities that I could loop
through and spit out. However, I see that
the way I was going about things was totally
not what you all were going for as I was trying
to make my code far more dynamic. I.e. I wasn't
hard coding the ingredients as you did. May be
helpful to future cohorts to lead them down
this path in some way...or not... :)